lib: Fix over-linking of events with identical IDs
The IDs of DflSources and DflMainContexts are based on the memory
addresses of the corresponding GSources and GMainContexts from the
recorded process. If a GSource or GMainContext is created, used, then
freed, that memory address may be re-used by the allocator, resulting in
the same ID appearing again in the logs. This is especially common with
GSources, which are created frequently.

This was resulting in events from subsequent instances of an object
being attached to the initial instance of the object, which was meaning
incorrect data was rendered in the timeline — for example, showing a
GSource being dispatched multiple times after being destroyed.

Fix this by introducing a new API in DflEventSequence for removing
groups of event sequence walkers once a final event is seen for them,
typically the free() call for the object instance. This has the added
benefit of tidying up some of the analysis code in dfl-source.c and
dfl-main-context.c.