lib: Skip to the right index when starting an iterator
If DflTimeSequence iterator is initialised with a non-zero timestamp,
use a binary search into the sequence to find the highest timestamp
index before the given timestamp (or the first index with an equal
timestamp to the given one) and start the iteration there. This can rude
the amount of iteration needed over large sequences if only a subset is
interesting.