build: Change library includes
Install the public headers so they are #included using a ‘lib’ prefix.
So now code should use:
   #include <libdunfell/dunfell.h>
rather than:
   #include <dunfell/dunfell.h>

This means the includes are the same for external and internal code,
which means the -I flags can be rationalised and we avoid some tricky
problems with internal consumers of the external header (dunfell.h).