lib: Fix corner cases in dfl_time_sequence_find_element()
This is used when initialising a DflTimeSequenceIter with a non-zero
timestamp. Fix corner cases around handling of subsequences of equal
timestamps in a DflTimeSequence — the iter should be set up to point to
the first of the set of equal-timestamp elements which are the greatest
of those with timestamps less than or equal to the given timestamp.